Time: 636 ms (85.22%), Space: 154.9 MB (24.13%) - LeetHub

+++ 2001-number-of-pairs-of-interchangeable-rectangles/2001-number-of-pairs-of-interchangeable-rectangles.cpp
@@ -6,7 +6,7 @@ public:
         }
         long long ans = 0;
         for(auto x:m)
-            ans+= ((long long)x.second*(x.second-1)/2);
+            ans+= x.second*(x.second-1)/2;
         return ans;
     }
 };
